requirements.txt
appdirs==1.4.3
asgiref==3.2.7
astroid==2.4.0
billiard==3.6.3.0
bs4==0.0.1
cachetools==4.1.0
camelot-py==0.8.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cmake==3.17.3
coala-utils==0.5.1
colorama==0.4.3
cssselect==1.1.0
distro==1.5.0
dlib==19.20.0
et-xmlfile==1.0.1
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.0.12
Flask==1.1.2
google==2.0.3
grpcio==1.30.0
gunicorn==20.0.4
idna==2.9
imutils==0.5.3
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
lazy-object-proxy==1.4.3
lxml==4.5.0
MarkupSafe==1.1.1
mccabe==0.6.1
mpmath==1.1.0
numpy==1.18.3
oauthlib==3.1.0
opencv-contrib-python==4.2.0.34
opencv-python==4.2.0.34
openpyxl==3.0.3
pandas==1.0.3
pdfminer.six==20200517
Pillow==7.1.2
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.7
pylint==2.5.0
PyPDF2==1.26.0
pyquery==1.4.1
PySocks==1.7.1
pytesseract==0.3.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
sortedcontainers==2.1.0
soupsieve==2.0
sqlparse==0.3.1
toml==0.10.0
urllib3==1.25.9
virtualenv==20.0.28
Werkzeug==1.0.1
wrapt==1.12.1
